INDI: Move copter lean vector attitude control upset recovery to m function
The reduced attitude control using projections of the lean vector
on the body-fixed x and y axis does not work in case the desired
thrust vector differs more than 90 degrees from the current thrust
vector. The modification to solve this problem was coded inside a
Matlab function block in the Simulink library model. This function
was cleaned up and moved to an m file. Moreover, in the Simulink
library file, both the m function leanVectorDerivTrafo and
indiCopterLeanVectorUpset are now used insteady of the code in the
Matlab function block.