Bug fix: wrong reduced attitude in case of NED and FRD frame are equal
If the attitude quaternion of the body-fixed frame (FRD) w.r.t.
the earth frame (NED) is perfectly [1;0;0;0], the yaw angle
of the quatReduced function was pi instead of 0. This caused
initialization problems of multicopters using this function for
control. The problem was solved by replacing the "sign" function
which returns 0 in case of input 0.