Copter aerodynamics: change definition of rate damping
The rate damping (relevant for rotor failures) is now
proportional to the angular rate (not to squared angular
rate) as in literature.
Moreover, it is proportional to the reference surface to
the power of 3/2 in the hope that this scales for quadrotors
of different size.
The default rate damping coefficient is calibrated for the
quadrotor Minnie (reaches 40rad/s in case of two opposing
rotor failures as in flight test)