Flexible body EoM: clean re-implementation of elastic equations of motion
Due to problems with aeroelastic stability in SE2A/ARGO, the elastic equations
of motion implementation was investigated again.
The following general problems were found:
- the node center of gravity is shifted, which was not considered
- the implementation was probably not consistent due to a mixture of the
  mean axis system and a fixed axis system
- some inertial coupling was considered, some was not
- the code was poorly documented
The problem was solved by a clean re-implementation according to the well-known
equations of Waszak & Schmidt. These equations are probably the most simple
ones since several effects are neglected like inertial coupling and varying
inertia. However, for a moderately flexible aircraft the equations should
still be suitable. If some effects should not be neglected, the implementation
would become much more complicated. This could probably be done in future
commits.
Moreover, the gravity forces are now applied in the node centers of gravity.