Bug fix: fuselage: compressibility correction was too high
The compressibility correction probably only needs to be applied
to the aerodynamic angles. However, it was also applied to the
width of the fuselage. That resulted in a stronger yawing moment
increase due to sideslip angles than the stabilizing yawing moment
increase due to sideslip of the vertical tailplane. This resulted
in a Mach number where the yaw rotation becomes unstable.
The problem was fixed by removing the compressibility correction
from the fuselage width.
Now the yawming moment dependency on the Mach number should be
the same as for the vertical tailplane.