VLM wing: Make reference point parameter of wing models more modular
The reference point (typically center of gravity) for the VLM
wing models was defined as distance from wing origin to the
reference point. That is prone to errors, because the wing
origin must be known outside of the VLM wing block although
it is a parameter that is known inside.
Since the wing origin is defined in c frame, the problem was
solved by defining the input reference position also in c frame.
The distance w.r.t. the wing origin is now computed inside the
block.