VLM wing and fuselage: Remove steps in unsteady fuselage forces and fix unsteady wing bugs
Fuselage:
If a 1-cos gust is simulated, there are steps in the fuselage
local aerodynamic coefficients as well as in the total forces.
Here, a temporary fix is applied in case of unsteady simulation.
The fix is achieved by not filtering the local aerodynamic angles
but the computed aerodynamic forces with a first order delay.
To do:
It should be generally reconsidered how to implement the local
fuselage aerodynamics.

VLM wing:
One main problem was that the unsteady configured wing produced a different
lift than a steady configured wing in case of steady state operation.
There were two bugs that fixed this problem:
1) The lift curve slope must not be corrected with the sweep angle in wingSetUnstAeroState
2) The local lift coefficient computed in wingSetLocalCoeff was slightly different
   from the previously computed lift coefficient in wingSetCirculationUnsteady
Additionally, the complicated circulation iteration in wingSetCirculationUnsteady
could be simplified. This was achieved by directly computing the corresponding
induced angle of attack in case of a steady configuration (this should converge faster).
In case of the unsteady configuration, no circulation iteration is required anymore.
Since the viscous lift coefficient won't change during the iterations, the viscous
lift coefficient is accepted to be the VLM lift coefficient. From this, the corresponding
steady state induced angle of attack can be obtained without iteration by computing
the VLM backwards.