Remove wing torsion flutter and acceleration unit delay feedback
The unsteady airfoil model needs the dimensionless pitch rate
as one of two inputs. The dimensionless pitch rate involves
the time derivative of the angle of attack which was computed
based on the time derivative of the local airspeed. The other
input was the angle of attack at 75% chord (control point
position). However, that is probably a mistake because it seems
that the angle of attack at 25% chord should be used [1].
Moreover, the dimensionless pitch rate is used instead of
the difference between 75% chord and 25% chord angle of attack,
see [1] Eq. between (18) and (19). Note that the current
implementation seems to cause unstable wing torsion at
low airspeed.
The problem was solved by replacing the 75% chord angle of attack
with the 25% chord angle of attack. Moreover, the dimensionless
pitch rate is now the difference between 75% chord and 25%
chord angle of attack (times 2). With this implementation the
time derivative of the local airspeeds is no longer needed.
Thanks to this, 3 unit delay blocks could be deleted: rigid
body acceleration, rigid body angular acceleration and structural
acceleration (all of them were no states so a unit delay was
required).
To do: the 25% chord angle of attack is currently only an
approximate computation.
To do: The wing torsion flutter seems to be disappeared.
However, the wing forward-backward bending still seems to be
unstable.

Literature:
[1] Leishman, J. G., & Nguyen, K. Q. (1990). State-space
    representation of unsteady airfoil behavior. AIAA journal,
    28(5),836-844.