Autopilot Dragonfly: change condition for loiter reference reset
In flight test the loiter flight mode did not work.
The quadcopter always started to fly upwards after
entering the controller. This problem could not be
observed in Simulation or SITL.
The guided flight mode was not tested yet in flight
test.
The condition for reseting the reference position
and velocity was probably dreadfully implemented
and may have caused the problem: the reference was
reseted in case two floats were not equal which is
probably not safe. This may have cause permanent
resets of the reference.
Now the condition is much simpler. If the number
of the flight mode is more than 0.5 away from the
number of the flight mode in the previous sample,
the reference will be reseted.