VLM wing: remove forward-backward wing bending flutter
Forward-backward wing bending flutter was observed.
It turned out that it came from the unsteady airfoil
aerodynamics. This linear state space model uses the
airspeed as a parameter. However, in the VLM wing
project the airspeed is allowed to vary. Unfortunately,
the unsteady airfoil model (which was implemented
according to literature) uses an unfavorable definition
of the state vector. The steady-state amplitude of
the state vector depends on the airspeed. This caused
a varying effective angle of attack in case of a
varying airspeed holding the geometry angle of attack
constant. And this caused that the lift vector was
rotated forward and backward when the wing was bending
forward and backward. This behavior is not desired!

The solution was to use a different definition of the
state vector. Now only the time constants depend on
the airspeed but the magnitude of the states does not
depend on the airspeed. Like this it is ok when the
airspeed changes. Moreover, in case of a fixed
airspeed and variable angle of attack the output of
the model is the same as before.