Unsteady fuselage aerodynamics: use local forces as states
In previous versions the aerodynamic angles were filtered
in case of unsteady aerodynamics. However, as the forces
are computed based on the differences of the aerodynamic
angles at different fuselage segments, the fuselage
forces used to step in case of gust penetration.
To avoid this problem, the aerodynamic angles are no
states anymore. In place of the aerodynamic angles the
aerodynamic forces are now filtered with the same time
constant.