VLM wing: improve unsteady flap lift distribution
During the comparison of the VLM wing with unsteady RANS
data for an unsteady flap deflection, the following
problems were observed:
- the local lift at the flap position was strongly delayed
- the local lift at the flap position did not change enough
The first problem was fixed by bypassing the spanwise downwash
around the 2d unsteady aerodynamics model (else the non-
circulatory lift fights against the lift change due to dynamic
flap deflection).
The second problem was fixed by removing the sweep compensation
for the flap effectiveness (this should be correct because
the lift curve slope also needs no sweep correction).
Moreover, the geometry generation was changed because the
distributed flap positions of the LEISA configuration
needed to be configured exactly. Before this commit, the
width of the wing panels was created first and then the flaps
were assigned to the panels (this changed the actual width of
the flaps). Now there is a panel border at each flap boarder
from the beginning and more panels are added afterwards until
the desired number of panels is reached.

(cherry picked from commit b2faa754b3788e7cc5f7bfdcf940a2f5b15b2c1c)