Make visualization of flexible unsteady aircraft nicer
Colors were changed:
- background: blue -> heaven blue
- forces: yellow -> green
- stall: green lines gradually become orange
- flaps: red -> grey
Surfaces were added. The surfaces are white and transparent.
Note that transparency render is a huge problem. If there
are multiple transparent surfaces in a row behind the camera,
the rendering is only correct if the objects with the
highest distance to the camera are rendered first.
That is why the order of rendering was changed at some places.
However, some unrealistic renderings remain.