VLM wing: Bug fix: wind non-circulatory lift direction of vtp
The non-circulatory lift due to wind always only
considered the wind component in aircraft z direction.
For the vertical tailplane that caused lateral forces
in case of vertical gusts.
The problem was solved by computing the local angles
of attack and dimensionless pitch rate perpendicular
to the inflow and the local surface normal vector.