Propeller map-based: Improve fit at low RPM
The curve fitting algorithm uses (absolute) least mean squares
as cost function. However, at low RPM the absolute thrust
and torque are very small compared to the maximum numbers
in the map. This easily leads to large "relative errors" at
low RPM.
The problem was solved by applying a weighting to the least
mean squares cost function which is large in case of values
close to zero and small in case of large values. Moreover,
since for some propeller maps there is a high variance at high
RPM (probably due to flow separation), a smoothing was applied
which reduces the variance.