Aeroelasticity: correct backward force transformation
For aeroelastic simulations the forces need to be
converted from structure nodes to aerodynamic nodes.
However, if a backward transformation is needed, this
transformation was not exact. Now this backward
transformation was corrected by using pinv insteady
of transpose. Moreover, a warning was added in case
the forces can not be transformed back accurately in
case of too many aerodynamic nodes.