EOM Flexible: consider node gravity moment due to cg offset
For the calculation of the loading of the structural nodes
by gravity, it was assumed that the center of gravity of
each node is located at the node position. In this way, no
moment due to gravity on the nodes was considered.
The problem was solved by considering the displacement of
the center of gravity for each node.