Battery: Revise and generalize discharge model
The battery discharge model of Tremblay uses absolute
parameter values like battery capacity, voltage,
internal resistance as well as hyper parameters that
depend on these absolute values.
etc. However, default battery parameters could be used
for a wide range of different batteries as long as
they are made of the same material (e.g. LiPo).
In this commit LiPo battery parameters are normalized
so that LiPo batteries of different capacities and
voltages can be simulated based on the same default
parameters.
Additionally, the code was revised and two Simulink
blocks were added.
Since the coupling of the battery model and the
"Brushless DC Motor" block in the library
actuators_lib.slx causes an algebraic loop, the
second battery block was introduced which can be
coupled with the "Brushless DC Motor".
In the actuators_lib.slx a 2nd order motor model was
added which can be coupled with the first battery
model without algebraic loop.
Note that the algebraic loop causes a warning but
neither decreases the simulation speed nor yields
wrong results.
To do: Improve documentation, add unit test

Ref. #12