Control: Replace "...LoadParams" functions with "loadParams" function
In the control folder there are several Simulink blocks that need
use a large amount of parameters. These parameters are mostly
stored inside a struct keep it clearly arranged. The parameters
are loaded using individual "...LoadParams" functions.
However, in most cases, the "...LoadParams" functions did nothing
else than loading the parameters from the parameter file.
That is why the new function "loadParams" was added which can
be used instead of the individual "...loadParams" functions.
The obsolete individual "...loadParams" functions were deleted.
In the Simulink models the mask description was adjusted.