INDI Copter: Add "Incremental specific thrust" library block
Moreover, it was discovered that the reference models were not
initialized correctly because of the internal discrete derivative
block (which are initialized with 0). The problem was solved
by making the derivatives additional inputs. Outside of the
blocks the signals should be filtered and the derivatives
should be initialized correctly.
On this occasion, the order of inputs has also been changed.