ArduPilot interface: Add measured values to ArduPilot interface
More signals were added to the measureBus (see initInterFaceBus).
Hence, the Simulink block "receive measureBus from ArduPilot SITL
had to be adjusted. initInterfaceBuses was converted to a function
because the struct variables are not needed in base workspace.
The sample time variable is not created anymore and has to be
set by the user.
The user can now specify (function input of initInterfaceBuses)
whether the interface variables are single or double precision.
Double precision caused cryptic variable names for the cmd and
measure bus after code generation.