Copter Pos Control: avoid copter flip after initialization
In the Simulink block "NDI position control for copters"
there is a low-pass filter for the reference signals.
Unfortunately, the states of the filters were initialized
with zeros, which means that the initial reference position
was zero. This caused violent reactions of the copter in
case the controller was initialized during flight.
The problem was solved by initializing the reference
position with the currently measured position.