ArduPilot interface: Update and simplify README
The README had to be updated because of the update
of the ArduPlane interface to version 4.3.0.
Moreover, it seemed to be to complicated to maintain the
provided git patches. That is why the README now refers
to the ArduPilot fork. With this the README becomes shorter
and hopefully more comprehensive.
The git patches were deleted.
The Simulink controller templates were moved upwards one
folder because they were the only remaining files in
their folder.