ArduPilot interface: Avoid data copying for logging
Previously, the output of the "log muxer" Simulink
block contained a buffer for the maximum of to be
logged signals plus parameters to indicate the labels
and signal names. Unfortunately in the generated C++
code all these outputs were copied each time the
step() function was called even if not needed.
Moreover, the logging could only be initialized after
calling the step() function in the generated C++ code
a first time which does not look clean.
Now the output of the "log muxer" Simulink block only
contains the signals to be logged. The parameters to
indicate the labels and signal names are now defined
by a "Simulink.Parameter" named "log_config" in the
Matlab base workspace. The "log_config" parameter is
created automatically by the mask initialization of the
"log muxer" block calling function ardupilotInitLogMuxer.
The "log_config" parameter is configured such that it
appears as initialized const parameter in the generated
C++ code (in MatlabController.cpp).