VLM wing: assure wind velocity (input) dimensions are correct
The wind velocity is an input of the high level function
wingSetState. It is intended to have the size 3xn_panel.
If something in Simulink wents wrong and the signal is
reshaped to a one-dimensional array, this one-dimensional
array is still accepted by the function because it only
requires the number of elements to be 3*n_panel.
If the reshaping is done such that the dimensions are
exchanged, weird behavior will be the result.
The problem was fixed by checking both dimensions in
wingSetState. We keep the "(:,:)" instead of deleting
it completely because else it would be possible to
change the dimensions in interpreter mode.