Fix trajFromWaypointsIterative for zero waypoints and improve speed
When the number of waypoint was zero, the function trajFromWaypointsIterative crashed cause of an array bound violation.

A new function trajValidateWaypoints is introduced to check if the given waypoints are enough for the requested type of trajectory (cycle option) and maybe fulfill some other conditions in the future, e.g. points are not equal.

For improving calculation time, the residuum of ||Ax-b|| is now calculated before the first iteration lsqr. If the residuum starts to grow, the lsqr iteration is finished an the loop can be breaked.