Flight mode AltHold: Speed up slow simulation
The simulation in LADAC-Examples including
the altitude hold flight mode was much slower
than with loiter flight mode.
This was caused by an algebraic loop. The
loop included the vertical acceleration
feedback.
The problem was solved by removing the initial
condition of the acceleration low-pass filter.
Moreover, the sensor filters are now linked
to the filters_lib.