Fix broken library browser
Due to the previous commit (change license
from GPL-2.0 to GPL-3.0) LADAC was broken
in the Simulink Library Browser.
It is not known why.
Solution:
All LADAC library files were re-saved by
the Simulink Library Browser (this
increases the file size btw).
Additionally, not in all Simulink library
files the license disclaimer was changed.
This was also fixed in this commit.