Rigid Body: Improve documentation and clean up
This revision had been planned for some time and has
now been carefully carried out.
The rigid body library is an essential part of many
models, so this update is likely to cause error
messages in projects that use LADAC as a submodule.
This is due to the fact that in the block "State
equations (Quaternions)" the bus signals have been
renamed (units have been removed).
Furthermore the block "State equations (Quaternions)
w/ non-constant g" has been deleted.
Instead of this block, the block "State equations
(Quaternions)" is to be used, in which an external
gravity can now be configured via the mask.
Finally a readme was added, which is linked in the
Simulink blocks and a test function was added.

The complete aircraft models (quadcopter, airplane,
tailsitter, tiltwing) were adjusted.

Note:
In the quadcopter library block the signal V_Kb
was passed to the ground model instead of V_Kg.
This bug was resolved en passant.

Ref. #15