Autopilot LindiCopter: Add stabilized flight mode
When stabilized flight mode is enabled, the
control allocation weighting of the vertical
acceleration is set to zero and the desired
control input is set to the throttle setting.
Moreover, the behavior of setting the motor
commands to idle in case of zero throttle
was removed because this behavior is not
expected by the pilot (this feature had been
added to improve landing - but this should
be done in a different way).
More details about Simulink files:
- ca_lib: In block "INDI high level wls
  control allocation for copters" the
  modification of "Delta gamma" was removed.
  This was replaced by setting the vertical
  acceleration weighting "Delta diag W_v(4)"
  to zero if input "is_manual_throttle" is
  true. The modification of "Delta u_d" was
  also changed.
- indi_lib: A new inport "is_manual_throttle"
  was added.
- lindiCopter_lib: In the subsystem "State
  Logic Bus" a bus is created that contains
  logical signals to switch block sources
  depending on the current flight mode.
  The subsystem "Select Atti Command Source"
  was added.
  The stick input commands are gathered in
  the stickCmdBus to make it cleaner.