Add new setpoint generator
The setpoint generator / reference model
'NDI position controller for copters reference model'
is now a variant subsystem and contains the 'old' PT1 based
method and a new calculated method called 'Generator'. This
new method calculates the setpoint signals based on the constraints.
It is more responsive than the PT1 method, but also almost always
pushes the acceleration to its maximum limits.

This commit also includes a bugfix for the 'old' PT1 method. The
velocity calculation for the xy-velocity did not linearly scale
the stick input to velocity commands.