Propulsion: Bug fix: wrong sign of propeller torque due to gyro precession
The sign of the propeller torque due to gyro precession
was wrong.
The bug was corrected by adding a "Gain" block with value
"-1" to the library block "Torque due to Gyro Precession".
This block is used inside the propulsion_lib library in the
block "Map-based propeller fixed to airframe".
Other affected libraries are:
- "Quadcopter Propulsion" in quadcopter_propulsion_lib
- "Quadcopter" in quadcopter_lib
- "Tailsitter Flight Dynamics" in tailsitter_lib
- "Tandem Tiltwing Dynamics" in tandem_tiltwing_lib
- "Conventional Airplane, single engine" in
  conventional_airplane_single_engine_lib