Autopilot LindiCopter: avoid unstable rotations after single rotor failure
In case of a single rotor failure the following
problem could be observed:
- In the transient phase from normal operation
  to relaxed hover solution (fast yaw rotations),
  it could happen that the quadrotor reaches a
  lean angle of circa 90 degrees and starts to
  rapidly rotate about the vertical earth axis.
  The quadrotor might not be able to recover.
  This was due to the fact that the quadrotor
  rotates circa about the desired thrust vector.
  The thrust vector was the control variable for
  reduced attitude control. In that case, the
  first and second time-derivative of the thrust
  vector are circa 0 although the quadrotor is
  rapidly spinning.
- If the quadrotor successfully reaches the
  relaxed hover solution, the quadrotor position
  could nicely be controlled as expected.

This commit is a first solution for the described
problem.
The problem was solved by linearly disabling the
original thrust vector controller from a desired
lean angle offset of 0 degrees to 90 degrees.
Simultaneously, a second controller is enabled
which regulates the pitch/roll rate and
acceleration in body-fixed axis. Note that the
second controller is not able to stabilize the
quadrotor position in the relaxed hover solution
but it helps to avoid the unstable rotations
described above.