Autopilot LindiCopter: fix stabilize flightmode
The stabilize flightmode (attitude controller)
did not work in ArduPilot SITL (same in flight
test). The problem was caused by the control
allocation module. The problem appeared if the
controller runs in single precision.
In the pure Simulink environment, however,
LindiCopter was run in double precision.
It has not been testet what happens when
LindiCopter runs in single precision in the
pure Simulink environment.
The problem was fixed by forcing the control
allocation module to run in double precision
even if the rest of the controller runs in
single precision. Moreover, a parameter was
added to disable forcing double precision.