Autopilot LindiCopter: fix stabilize flightmode
The stabilize flightmode still did not work
in flight test. The problem was solved by
reducing the control allocation weighting
of roll and pitch acceleration from 30 to 1
in case of manual throttle (this is done in
ca_lib.slx.
The forcing of double precision was no
longer needed and removed.
The stabilize flight mode of this
implementation has been successfully tested
in flight test.