Joystick: Fix calibration for new Matlab versions
The joystick calibration was successfully tested
with Matlab R2018b. Using Matlab R2023a there was
an error:
Error: File: joystickCalibrate.m Line: 104 Column: 35
Brace indexing into the result of a function call is
not supported. Assign the result of 'joystickOut' to
a variable first, then brace index into it.

Matlab R2023a throws the error because it detects a
syntax error: The variable joystickOut was not defined
in the code. However, the code would run because
joystickOut will be created in the command
sim('joystick_setup').
The problem was solved by initializing joystickOut
as empty cell array.