ArduPilot SITL interfaces for Simulink: README add workaround for UDP connection problems when using WSL2
Workaround added for a connection problem when ArduPilot SITL is used under WSL while MATLAB/Simulink runs under Windows. The Windows Subsystem for Linux (WSL2) does not fully support UDP port forwarding which results in Simulink not being able to establish a connection to ArduPilot SITL and vice versa. The solution to this problem is the tool socat, with which UDP packages can be sent via a TCP tunnel.