Replace "Qt_OpenGL_visualization" with "FlexiFlightVis"
The aeroelastic flight dynamics simulation visualization
program "Qt_OpenGL_visualization", written in C++ with
Qt, was located in LADAC. The "Qt_OpenGL_visualization"
source code had to be compiled by the user. The
compilation required installation of Qt und some
dependencies.
To simplify that, the "Qt_OpenGL_visualization" source
code was moved to a separate Repository and named
FlexiFlightVis:
https://github.com/iff-gsc/FlexiFlightVis
In that repository there is a Readme and an automatred
Github workflow to provided an AppImage for Linux. In
this way, the user only has to download the AppImage
instead of compiling the code locally.
In LADAC an improved example was created.