ArduPilot custom controller: Rename a_Kb, a_Kg -> a_b, a_g
The choice of the variable names a_Kb and a_Kg were
unfavorable. The variables are the acceleration in the
body-fixed coordinate system b and earth-fixed coordinate
system g. The index K indicates a motion from the body
relative to the earth. However, the variables a_Kg and
a_Kb are defined to represent the acceleration measured
by the IMU and thus contain the gravitational acceleration.
To avoid confusion in the future, the "K"s in the variable
names have been removed. Moreover, the comments in the
functions were supplemented, so that the definition of the
variables becomes clear.
Unfortunately, this change will cause errors in old
projects.