INDI CA: Add rate limit constraint
You may want to define a rate limit constraint
if the used actuator has a rate limit.
This is now possible by a new input of the
function caIndiWls. This input is added as an
optional input of the Simulink block "INDI wls
control allocation" in ca_lib.slx.
Since the Simulink block now has several
optional inputs, it probably makes more sense
to make these inputs also optinal in the
function caIndiWls. This was achieved by now
grouping the optional inputs with varargin
and specifying them as name-value arguments.