Airplane aerodynamics: Fix VTP position
There was a yawing moment due to drag at
the vertical tailplane (VTP). It was found
that the vertical tailplane was shifted
laterally instead of vertically w.r.t. the
body frame.
The error was fixed by applying the VTP
rotation matrix in the 2nd and 3rd mask
parameter of the block "wing 2 point
aerodynamics (vertical tailplane, vtp)".