Remove small error in mask for discrete PT2 in filters_lib
In the new mask of the discrete PT2 block an empty text field
was used as a spacer. However, it seems that Matlab adds
'(Add text here)' to an empty text field each time the mask is
opened for editing. Therefore the text field was replaced with an
empty panel.