Fix position time constant calculation
When calculating psc.rm.veltc, only the physical time constant was considered.
Now the calculation also takes into account the time constant of the separation
factor and uses the larger value.

This means that the results for psc.rm.veltc and for the position controller
gains (psc.k.<...>) may be different than before.