ArduPilot custom controller (Plane): Add servo function "passthrough"
The ArduPlane interface between the Matlab/Simulink controller
and the ArduPlane channel outputs is based on the ArduPlane
servo function. (The ArduCopter interface uses the channel
number and does not contain the ArduCopter servo functions)
It was not possible to bypass/ignore the ArduPlane servo
functions, which is a limitation. Moreover, the interface was
limited to 8 channels.
That is why in the Simulink block "Actuator Muxer ArduPlane"
in ardupilot_custom_controller_lib.slx the servo function
"feedthrough" was added. A mask parameter was added to
specify the the number of channels (1-16).
The initialization of the block was moved to the new
function ardupilotPlaneActuatorMuxerBlockInit.m and the
ArduPlane servo function list was moved to the new file
ardupilot_plane_servo_functions.txt to make the interface
easier to maintain.