Fix maximum velocity calculation
The calculation of the maximum climbing velocity (velumax) did not
take the earth's gravitational force into account and led to an
overestimated result. In addition, the aggr_pos factor was applied
twice, firstly for the motor setpoint and secondly for scaling
the calculated velocity. The motor setpoint also did not take into
account the maximum allowed setpoint (u_max). Furthermore, it was
assumed that the maximum horizontal velocity (velxymax) is equal to
the maximum climbing velocity, which does not consider a possible
lean angle limitation.

To solve these problems, the calculation of the maximum velocity was
split into two separate calculations, one for maximum climbing velocity
(velumax) and one for the maximum horizontal velocity (velxymax). The
motor setpoint is now based on a scaled down version of the maximum
motor setpoint (u_max) and the calculation of the horizontal velocity
considers the lean angle limitation.