Fix maximum rotational acceleration calculation
The maximum rotational acceleration was only calculated for the roll
and pitch axes and the smallest value was used as the limiting value.
For a quadrocopter, however, the limiting case is the "diagonal" case,
in which only two propellers are used for the rotational acceleration.
The new formula is a simplified approach to calculating the reduced
diagonal acceleration.