Unsteady flap: Bug fix: Fix compressibility correction
The comparison of the model for unsteady flap aerodynamics
with CFD data has shown that the steady-state pitching moment
is clearly too high at high Mach numbers.
The error lay in the compressibility correction.
The pitching moment was increased instead of decreased due to
high Mach numbers.
The model for steady-state flap aerodynamics was correctly
implemented.
The documentation of the functions was updated for both the
unsteady and steady-state models as it was outdated.