Remove matrices from LindiCopter parameters struct
Converted matrices to arrays in the control allocation field (ca) of
the LindiCopter parameters struct. This is required to be compatible
with the ArduPilot parameter ecosystem.

This is a breaking change that introduces incompatibility with the old
LindiCopter parameters struct format. This affects the functions
'caIndiWls.m' and 'caWlsCopterUpset' and the Simulink library ca_lib,
especially the block 'INDI wls control allocation', which is probably also
used in other projects.