Improve wrong input handling in ca_lib
In 954a90ad, the format of the control allocation struct (ca) was
changed and an input check was added to 'ca_lib/INDI wls control
allocation' in the initialization of the mask. The input check did not
cancel the full model update and many errors were shown with the wrong
input format. This has been changed to improve the interpretation of
the error.