VLM wing: Add function to get aerodynamic derivatives
The function wingGetDerivs.m was added to obtain the
aerodynamic derivatives of a vlm wing struct.
The script wing_parameteric.m was adjusted such that
optionally flaps can be specified. The script can be
used to validate the derivatives of simple wing geometries.
The flaps were added to check the derivatives w.r.t. flap
deflections.
Some bugs had to be fixed for the code to work with
non-symmetrical wings.