Autopilot LindiPlane: Remove control effectiveness for input derivatives
The control effectiveness did not only consider the
flap deflections but also the deflection rate (not
implemented) as well as the deflection acceleration.
The influence of the deflection rate was not
implemented and the influence of the deflection rate
was disabled by default by setting the flap inertia
to zero. These effect have not been used and seem to
be negligible.
Moreover, the consideration of these effects in the
code uses a relatively high amount of memory.
In order to reduce the memory consumption, the
consideration of these effects are removed.