Simple wing: Bug fix: Consider flap moments properly
There was a bug in the Simulink block (Matlab function)
simple_wing_lib/Simple Wing Aerodynamics/moment coefficients/
Sum moment coefficients.
The code should be the same as in the pure Matlab implementation
in function simpleWingRun in line 88-89. However, in the
Simulink implementation the variable C_lmn_flap_np_w was not
considered (overwritten).
This variable contains the flap moments with respect to the
wing neutral point. Thus, before the bug fix, all the flap
forces were applied at the wing neutral point ignoring the
individual centers of pressure of the flaps.