Autopilot LindiPlane: Fix gyro low-pass filter
For the gyros a double 2nd order low-pass filter is
used in order to strongly reduce high-frequency noise.
Both of these filters use the duplicate cutoff
frequency (2*ap.sflt.omega) so that the overall cutoff
frequency approximately matches that of a single 2nd
order low-pass filter.
In Commit 52a6ea28 a bug was introduced. Instead of the
duplicate cutoff frequency the unmodified cutoff
frequency was used leading to a too high signal delay
and consequently decreasing the control performance.
This bug was fixed by this commit.