Autopilot LindiPlane: Improve inversion quality by filter update
For the attitude feedback controller, the desired values are
obtained by filtering the reference values from the reference
model. The filteres were 2nd order filters summarizing the
delay due to actuator dynamics (2nd order) and sensor filter
(2nd order). This 2nd order filter was replaced by two 2nd
order filters. With this modification, the inversion of the
roll acceleration is perfect (if lindi.sflt.numGyrFilt = 1
and lindi.mla.use = 0).
The inversion of the pitch rate is satisfactory.