Simple airplane aerodynamics: Add wing downwash
The Simulink block "Conventional Airplane Aerodynamics simple"
in "conventional_airplane_aerodynamics_lib" now contains
downwash of the main wing on the horizontal tailplane.
Therefore new "downwash" parameters are added to the parameters
struct. The parameters are computed from new functions using
lifting line theory. In case the new parameters are missing
from the parameters file, the downwash parameters are set to
zero when loading the parameters with
"conventionalAirplaneAeroLoadParams". That is why this update
should be backward compatible.