Autopilot LindiPlane: Remove control allocation iterations
Control allocation iterations occur as soon as one or
more flaps are in saturation. In this case other flaps
are actuated to increase the flap forces and moments.
However, this can be quite costly and may be undesired
when the code runs on a micro controller. That is why
control allocation iteration is deactivated by default.
Moreover, a lower priority is given to the elevator so
that the main wing flaps are used less for pitching.