Autopilot LindiPlane: Add faster direct lift control option
The control allocation of 3x rotational accelerations and
vertical acceleration are now separated. This enables the
possibility to use a faster low-pass filter for vertical
acceleration, which increases performance of gust rejection.
For feedback of the 3x rotational accelerations, typically,
the measured gyro signals must be derived w.r.t. time, which
typically requires the use of a (relatively slow) low-pass
filter to avoid noise. For vertical acceleration, the measured
signal does not need to be derived w.r.t. time, which is why
a faster low-pass filter can typically be used.
The low-pass filter for vertical acceleration is defined by the
new parameters struct ap.dlc.flt. It is enabled when setting
ap.dlc.opt=3.
In the Simulink model the blocks "Measurement Filtering",
"Direct Lift Control", and "Plane Inner Loop INDI and Control
Allocation" were changed.