Autopilot LindiPlane: Fix DLC for dlc.opt=2
In the last commit about LindiPlane (Commit 65dddb00)
a faster direct lift control option (dlc.opt=3) was
added. However, an error was introduced for standard
direct lift control (dlc.opt=2). In case of dlc.opt=2
the same low-pass filter should used for the vertical
acceleration feedback as for rotational feedback
(sflt). But in the block "Plane Inner Loop INDI and
Control Allocation/Direct lift control - control
allocation" the DLC-specific filter (dlc.flt) was
used. This can lead to oscillations in case the
filters are parameterized differently.
In this commit, this filter block was replaced by a
block that is linked to the feedback filter block
so that always the same filter is used.